Symbol detangling
We can not inhertit, we need unique symbols.
