+++ include/openPMD/IO/ADIOS/ADIOS1IOHandler.hpp
@@ -22,17 +22,11 @@
 
 #include "openPMD/IO/AbstractIOHandler.hpp"
 
-#if openPMD_HAVE_ADIOS1
-#   include "openPMD/IO/AbstractIOHandlerImpl.hpp"
-#   include <adios_read.h>
-#endif
-
 #include <future>
 #include <memory>
 #include <string>
 #if openPMD_HAVE_ADIOS1
-#   include <unordered_map>
-#   include <unordered_set>
+#   include <queue>
 #endif
 
 #if _MSC_VER
@@ -44,86 +38,40 @@
 
 namespace openPMD
 {
+    class EXPORT ADIOS1IOHandlerImpl;
+
 #if openPMD_HAVE_ADIOS1
-class EXPORT ADIOS1IOHandler;
+    class EXPORT ADIOS1IOHandler : public AbstractIOHandler
+    {
+        friend class ADIOS1IOHandlerImpl;
 
-class EXPORT ADIOS1IOHandlerImpl : public AbstractIOHandlerImpl
-{
-public:
-    ADIOS1IOHandlerImpl(AbstractIOHandler*);
-    virtual ~ADIOS1IOHandlerImpl();
+    public:
+        ADIOS1IOHandler(std::string const& path, AccessType);
+        virtual ~ADIOS1IOHandler() override;
 
-    virtual void init();
+        virtual std::future< void > flush() override;
 
-    virtual std::future< void > flush() override;
+        virtual void enqueue(IOTask const&) override;
 
-    virtual void createFile(Writable*, Parameter< Operation::CREATE_FILE > const&) override;
-    virtual void createPath(Writable*, Parameter< Operation::CREATE_PATH > const&) override;
-    virtual void createDataset(Writable*, Parameter< Operation::CREATE_DATASET > const&) override;
-    virtual void extendDataset(Writable*, Parameter< Operation::EXTEND_DATASET > const&) override;
-    virtual void openFile(Writable*, Parameter< Operation::OPEN_FILE > const&) override;
-    virtual void openPath(Writable*, Parameter< Operation::OPEN_PATH > const&) override;
-    virtual void openDataset(Writable*, Parameter< Operation::OPEN_DATASET > &) override;
-    virtual void deleteFile(Writable*, Parameter< Operation::DELETE_FILE > const&) override;
-    virtual void deletePath(Writable*, Parameter< Operation::DELETE_PATH > const&) override;
-    virtual void deleteDataset(Writable*, Parameter< Operation::DELETE_DATASET > const&) override;
-    virtual void deleteAttribute(Writable*, Parameter< Operation::DELETE_ATT > const&) override;
-    virtual void writeDataset(Writable*, Parameter< Operation::WRITE_DATASET > const&) override;
-    virtual void writeAttribute(Writable*, Parameter< Operation::WRITE_ATT > const&) override;
-    virtual void readDataset(Writable*, Parameter< Operation::READ_DATASET > &) override;
-    virtual void readAttribute(Writable*, Parameter< Operation::READ_ATT > &) override;
-    virtual void listPaths(Writable*, Parameter< Operation::LIST_PATHS > &) override;
-    virtual void listDatasets(Writable*, Parameter< Operation::LIST_DATASETS > &) override;
-    virtual void listAttributes(Writable*, Parameter< Operation::LIST_ATTS > &) override;
-
-    virtual int64_t open_write(Writable *);
-    virtual ADIOS_FILE* open_read(std::string const& name);
-    void close(int64_t);
-    void close(ADIOS_FILE*);
-
-protected:
-    int64_t m_group;
-    std::string m_groupName;
-    ADIOS_READ_METHOD m_readMethod;
-    std::unordered_map< Writable*, std::shared_ptr< std::string > > m_filePaths;
-    std::unordered_map< std::shared_ptr< std::string >, bool > m_existsOnDisk;
-    std::unordered_map< std::shared_ptr< std::string >, int64_t > m_openWriteFileHandles;
-    std::unordered_map< std::shared_ptr< std::string >, ADIOS_FILE* > m_openReadFileHandles;
-    std::unordered_map< ADIOS_FILE*, std::vector< ADIOS_SELECTION* > > m_scheduledReads;
-};  //ADIOS1IOHandlerImpl
-
-class EXPORT ADIOS1IOHandler : public AbstractIOHandler
-{
-    friend class ADIOS1IOHandlerImpl;
-
-public:
-    ADIOS1IOHandler(std::string const& path, AccessType);
-    virtual ~ADIOS1IOHandler() override;
-
-    virtual std::future< void > flush() override;
-
-    virtual void enqueue(IOTask const&) override;
-
-private:
-    std::queue< IOTask > m_setup;
-    std::unique_ptr< ADIOS1IOHandlerImpl > m_impl;
-};  //ADIOS1IOHandler
+    private:
+        std::queue< IOTask > m_setup;
+        std::unique_ptr< ADIOS1IOHandlerImpl > m_impl;
+    }; // ADIOS1IOHandler
 #else
-class EXPORT ADIOS1IOHandlerImpl
-{ };
+    class EXPORT ADIOS1IOHandler : public AbstractIOHandler
+    {
+        friend class ADIOS1IOHandlerImpl;
 
-class EXPORT ADIOS1IOHandler : public AbstractIOHandler
-{
-    friend class ADIOS1IOHandlerImpl;
+    public:
+        ADIOS1IOHandler(std::string const& path, AccessType);
+        virtual ~ADIOS1IOHandler() override;
 
-public:
-    ADIOS1IOHandler(std::string const& path, AccessType);
-    virtual ~ADIOS1IOHandler() override;
+        virtual std::future< void > flush() override;
 
-    virtual std::future< void > flush() override;
-
-private:
-    std::unique_ptr< ADIOS1IOHandlerImpl > m_impl;
-};  //ADIOS1IOHandler
+    private:
+        std::unique_ptr< ADIOS1IOHandlerImpl > m_impl;
+    }; // ADIOS1IOHandler
 #endif
 } // openPMD
+
+#undef EXPORT
